[Chetan]Add. Retrieving data from the database.

+++ EmployeePayrollServices/EmployeeModel.cs
@@ -13,10 +13,10 @@ namespace EmployeePayrollServices
         public string EmployeeAddress { get; set; }
         public string Department { get; set; }
         public string PhoneNumber { get; set; }
-        public double Basic_Pay { get; set; }
-        public double Deductions { get; set; }
-        public double Taxable_Pay { get; set; }
-        public double Income_Tax { get; set; }
-        public double Net_Pay { get; set; }
+        public decimal Basic_Pay { get; set; }
+        public decimal Deductions { get; set; }
+        public decimal Taxable_Pay { get; set; }
+        public decimal Income_Tax { get; set; }
+        public decimal Net_Pay { get; set; }
     }
 }
